Move deposit wizard view/data helpers to the presenter
Move the ~25 config/state/form/stepper/admin-set/file/review helpers off
the Context concern and onto Hyku::DepositWizard::Presenter, so the logic
lives on one object instead of controller helper methods. The Context
concern now just delegates those names to the shared presenter instance;
views still reach them through that bridge for now. No behavior changes.
