+++ app/controllers/concerns/hyrax/deposit_wizard/navigation.rb
@@ -71,12 +71,12 @@ module Hyrax
       # strings/arrays) are stored so they serialize into the session.
       def advance_from_details
         build_work_form
-        if @form.validate(work_params)
+        if work_form.validate(work_params)
           wizard_state.attributes = work_params.to_unsafe_h
           next_step = wizard_state.uploaded_file_ids.any? ? 'file_meta' : 'review'
           redirect_to main_app.deposit_wizard_step_path(step: next_step)
         else
-          flash_error('hyku.deposit_wizard.errors.details_invalid', form_error_messages(@form))
+          flash_error('hyku.deposit_wizard.errors.details_invalid', form_error_messages(work_form))
           render :details
         end
       end
